Preparations for 0.17.0.
- Version number update.
- Cleaned up SVG files.
- Some tweaks to Play to make large carriable objects get carried better.
- PNG crush'd the new PNG files.
- Cleaned up long lines in the code.
- Implemented some -Wall and cppcheck suggestions.
- Corrections for Visual Studio. It wasn't including std::vector in particle.h, and std::unordered_set doesn't have .reserve(). For some reason.
- Visual Studio project files updated.

+++ Source/source/mob_fsms/pile_fsm.cpp
@@ -75,7 +75,7 @@ void pile_fsm::be_attacked(mob* m, void* info1, void* info2) {
     resource* resource_to_pick_up = NULL;
     pikmin* pikmin_to_start_carrying = NULL;
     
-    for(size_t r = 0; r < amount_to_spawn; ++r) {
+    for(size_t r = 0; r < (size_t) amount_to_spawn; ++r) {
         point spawn_pos;
         float spawn_z = 0;
         float spawn_angle = 0;
